Drop Cwt requirement
