+++ s4torch/model.py
@@ -10,7 +10,7 @@ from typing import Optional, Type
 import torch
 from torch import nn
 
-from s4torch._encoders import StandardEncoder, WaveletEncoder
+from s4torch.aux.encoders import StandardEncoder
 from s4torch.block import S4Block
 from s4torch.dsp.utils import next_pow2
 
@@ -27,9 +27,9 @@ def _parse_pool_kernel(pool_kernel: Optional[int | tuple[int]]) -> int:
 
 
 def _seq_length_schedule(
-    n_blocks: int,
-    l_max: int,
-    pool_kernel: Optional[int | tuple[int]],
+        n_blocks: int,
+        l_max: int,
+        pool_kernel: Optional[int | tuple[int]],
 ) -> list[tuple[int, int]]:
     ppk = _parse_pool_kernel(pool_kernel)
 
@@ -77,20 +77,20 @@ class S4Model(nn.Module):
     """
 
     def __init__(
-        self,
-        d_input: int,
-        d_model: int,
-        d_output: int,
-        n_blocks: int,
-        n: int,
-        l_max: int,
-        wavelet_tform: bool = False,
-        collapse: bool = False,
-        p_dropout: float = 0.0,
-        activation: Type[nn.Module] = nn.GELU,
-        norm_type: Optional[str] = "layer",
-        norm_strategy: str = "post",
-        pooling: Optional[nn.AvgPool1d | nn.MaxPool1d] = None,
+            self,
+            d_input: int,
+            d_model: int,
+            d_output: int,
+            n_blocks: int,
+            n: int,
+            l_max: int,
+            wavelet_tform: bool = False,
+            collapse: bool = False,
+            p_dropout: float = 0.0,
+            activation: Type[nn.Module] = nn.GELU,
+            norm_type: Optional[str] = "layer",
+            norm_strategy: str = "post",
+            pooling: Optional[nn.AvgPool1d | nn.MaxPool1d] = None,
     ) -> None:
         super().__init__()
         self.d_input = d_input
@@ -113,9 +113,9 @@ class S4Model(nn.Module):
         )
 
         if wavelet_tform:
-            from s4torch.dsp.cwt import Cwt
+            from s4torch.dsp.cwt import Cwt, CwtWithAdapter
 
-            self.encoder = WaveletEncoder(
+            self.encoder = CwtWithAdapter(
                 Cwt(next_pow2(self.l_max)),
                 d_model=self.d_model,
             )
